% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidlr.R
\name{fidlr}
\alias{fidlr}
\title{fidlr: FInancial Data LoadeR}
\usage{
fidlr()
}
\value{
Return either a csv file per instrument or a data frame. The csv file is stored in the working directory and the data frame in the globalenv.
}
\description{
fidlr is an RStudio addin designed to simplify the financial data downloading process.
This initial version is a wrapper around the getSymbols function in the quantmod package and only Yahoo, Google, FRED and Oanda are supported.
More data providers and functionalities might be added over time.
}
\author{
Arnaud Amsellem
}
\seealso{
\code{quantmod}

\code{shiny}
}

